requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.2.3
fastapi==0.96.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
pydantic==1.10.8
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
unificontrol==0.2.9
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchgod==0.7
websockets==10.1
yarl==1.9.2
